fix(EngineKit): simulatedWorkNanoseconds as Int, not UInt64 (symbol-graph violation)
RP relayed the macOS symbol-graph check: Swift.UInt64 appeared in EngineKit's
public surface outside allowed-types/EngineKit.json. Root cause: the three
fakes' simulatedWorkNanoseconds convenience property (my own test-fixture
design, not named by C-011/C-012) was typed UInt64 to match Task.sleep's
parameter directly. Changed the public type to Int (already allowed) and
convert at the Task.sleep call site.

+++ Packages/Core/Sources/EngineKit/Fakes/FakeEngines.swift
@@ -15,7 +15,7 @@ import Foundation
 
 public final class FakeTranscriptionEngine: TranscriptionEngine, @unchecked Sendable {
     public let engineId: String
-    public var simulatedWorkNanoseconds: UInt64 = 1_000_000
+    public var simulatedWorkNanoseconds: Int = 1_000_000
     public var modelVersion = "fake-1.0"
     /// К1: тест подставляет результат, чей throwing-init заведомо бросает
     /// `DomainValidationError` — движок обязан поймать её и завернуть в `.invalidResult`.
@@ -55,7 +55,7 @@ public final class FakeTranscriptionEngine: TranscriptionEngine, @unchecked Send
             progress(.started(stage: .asr))
         }
         do {
-            try await Task.sleep(nanoseconds: simulatedWorkNanoseconds)
+            try await Task.sleep(nanoseconds: UInt64(simulatedWorkNanoseconds))
         } catch {
             throw EngineError.cancelled
         }
@@ -107,7 +107,7 @@ public final class FakeTranscriptionEngine: TranscriptionEngine, @unchecked Send
 
 public final class FakeDiarizationEngine: DiarizationEngine, @unchecked Sendable {
     public let engineId: String
-    public var simulatedWorkNanoseconds: UInt64 = 1_000_000
+    public var simulatedWorkNanoseconds: Int = 1_000_000
     public var modelVersion = "fake-1.0"
     public var forcedResult: (() throws -> DiarizationResult)?
 
@@ -125,7 +125,7 @@ public final class FakeDiarizationEngine: DiarizationEngine, @unchecked Sendable
         }
         progress(.started(stage: .diarization))
         do {
-            try await Task.sleep(nanoseconds: simulatedWorkNanoseconds)
+            try await Task.sleep(nanoseconds: UInt64(simulatedWorkNanoseconds))
         } catch {
             throw EngineError.cancelled
         }
@@ -174,7 +174,7 @@ public final class FakeEmbeddingEngine: EmbeddingEngine, @unchecked Sendable {
 
 public final class FakePostProcessor: PostProcessor, @unchecked Sendable {
     public let engineId: String
-    public var simulatedWorkNanoseconds: UInt64 = 1_000_000
+    public var simulatedWorkNanoseconds: Int = 1_000_000
     public var promptVersion = "fake-1.0"
     public var forcedResult: (() throws -> [MeetingOutputDraft])?
 
@@ -187,7 +187,7 @@ public final class FakePostProcessor: PostProcessor, @unchecked Sendable {
     ) async throws -> [MeetingOutputDraft] {
         progress(.started(stage: .postProcess))
         do {
-            try await Task.sleep(nanoseconds: simulatedWorkNanoseconds)
+            try await Task.sleep(nanoseconds: UInt64(simulatedWorkNanoseconds))
         } catch {
             throw EngineError.cancelled
         }
